==> scripts/01-download_data.R <==
# Explain how download the data

#The General Social Survey (GSS) is a nationally representative survey conducted in the United States since 1972 to 2012. 
# It gathers information on various social and demographic trends and attitudes, including residents' living conditions, working status, 
#happiness, and many other variables of interest. To access and analyze the GSS data, one can download the data package from the GSS website. 
#The data are typically provided in a statistical software format, such as STATA or SPSS, and may require some data cleaning and processing before analysis.
#For this specific article, the author plans to simulate and clean the data to select the variables that are relevant to the research question. This may involve. 
#dropping irrelevant variables, transforming, or recording variables, and creating new variables from existing ones. Once the data cleaning is complete, the author can
#start analyzing the selected variables to investigate the research question. This may involve using statistical software or other analytically tools to explore relationships 
#and patterns in the data. Overall, the GSS data set is a rich source of social and demographic data that can provide valuable insights into various trends and attitudes over time. 
# However, careful attention to data cleaning and analysis is necessary to ensure the validity and reliability of the findings.
